@@ -1,4 +1,4 @@
-const mongoose = require("mongoose");
+import mongoose from "mongoose";
 
 const MessageSchema = mongoose.Schema(
   {
@@ -17,6 +17,6 @@ const MessageSchema = mongoose.Schema(
   }
 );
 
-const Message = mongoose.model("Messages", MessageSchema);
+const Messages = mongoose.model("Messages", MessageSchema);
 
-export default Message;
+export default Messages;
