@@ -1,6 +1,8 @@
-const Messages = require("../models/messageModel");
+// const Messages = require("../models/messageModel");
 
-module.exports.getMessages = async (req, res, next) => {
+import Messages from "../models/Message.js";
+
+const getMessages = async (req, res, next) => {
   try {
     const { from, to } = req.body;
 
@@ -22,7 +24,7 @@ module.exports.getMessages = async (req, res, next) => {
   }
 };
 
-module.exports.addMessage = async (req, res, next) => {
+const addMessage = async (req, res, next) => {
   try {
     const { from, to, message } = req.body;
     const data = await Messages.create({
@@ -37,3 +39,5 @@ module.exports.addMessage = async (req, res, next) => {
     next(ex);
   }
 };
+
+export { getMessages, addMessage };
