@@ -6,6 +6,7 @@ import { fileURLToPath } from "url";
 import mongoConnect from "./config/db.js";
 import { errorResponserHandler } from "./middleware/errorHandler.js";
 import { invalidPathHandler } from "./middleware/errorHandler.js";
+import { Server } from "socket.io";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
 // const __dirname = path.resolve();
@@ -17,6 +18,7 @@ import cors from "cors";
 import userRoutes from "./routes/userRoutes.js";
 import postRoutes from "./routes/postRoutes.js";
 import commentRoutes from "./routes/commentRoutes.js";
+import messageRoutes from "./routes/messageRoutes.js";
 dotenv.config();
 mongoConnect();
 const app = express();
@@ -30,6 +32,7 @@ app.get("/", (req, res) => {
 app.use("/api/users", userRoutes);
 app.use("/api/posts", postRoutes);
 app.use("/api/comments", commentRoutes);
+app.use("/api/messages", messageRoutes);
 
 //static assets
 app.use("/uploads", express.static(path.join(__dirname, "/uploads")));
@@ -38,4 +41,28 @@ app.use(invalidPathHandler);
 app.use(errorResponserHandler);
 
 const PORT = process.env.PORT || 5000;
-app.listen(PORT, () => console.log(`server is running on port ${PORT}`));
+const server = app.listen(PORT, () =>
+  console.log(`server is running on port ${PORT}`)
+);
+
+const io = new Server(server, {
+  cors: {
+    origin: "http://localhost:8000",
+    credentials: true,
+  },
+});
+
+global.onlineUsers = new Map();
+io.on("connection", (socket) => {
+  global.chatSocket = socket;
+  socket.on("add-user", (userId) => {
+    onlineUsers.set(userId, socket.id);
+  });
+
+  socket.on("send-msg", (data) => {
+    const sendUserSocket = onlineUsers.get(data.to);
+    if (sendUserSocket) {
+      socket.to(sendUserSocket).emit("msg-recieve", data.msg);
+    }
+  });
+});
